Fix info area overdraw with shapes
It seems that using SimpleScaleShape instead of the original MegaSimpleScaleShape causes the width to decrease by 1 and the height to increase by 1. I've no idea why, but the original MegaSimpleScaleShape is a terrible mess, and I don't feel like messing with it right now. Just bear in mind that there may be further issues caused by the change of functions.