Cleanup 3d_inter.c
This is a pretty small file, but there's a bunch of stuff in 3d_agent.c that really belongs here.

Since we no longer preload graphics from the vswap, there's not much point to the PreloadGraphics function anymore, so now it's just DrawLevelTransition. There might be a need to revisit it later with the LS_ thermometer variables, however.