Add main menu code
This is mostly the same as in Wolf4SDL, with several new functions.

The font used to display the control strings doesn't have any lowercase characters, so the ScanNames array MUST be uppercase only. However, this could be safeguarded against by using toupper on any strings passed to the printing routines.