Implement movie animations
This was surprisingly easy to port, considering how much DOS/VGA fluff there was to deal with. I ended up ripping out all the memory block caching & latch updating and replaced it with a single memory buffer that loads the whole anim file and a simple function that draws a segment of pixels at the given screen offset.

I've only tested it with the intro animation so far, but that one seems to work fine at least.