Cleanup the actor code
Here they are at last, the actor states and functions. Note that this is NOT fully tested and is NOT working properly yet!

Aside from other issues, the most glaring issue is 16 bit objstruct members being used to hold memory addresses, which was a clever hack to save memory in DOS, but is quite honestly a terrible idea. objstruct should be given pointers to hold those addresses directly, or at the very least those members should hold indexes into the objlists for the object they want to store instead of the memory address.