Fix obj->dir wrapping around
This wasn't a problem in the original code, probably because the dir member uses the dirtype typedef which I imagine is signed short in DOS. But now it's an unsigned byte, so we need to be more careful.

Also changed the weird double bitshift to test if obj->dir is odd - it's more straightforward to just AND it with 1.