Fix cloaked shape intensity
This is a weird part of the original game where I've been unable to properly replicate the behavior. Although this is what the code SHOULD be doing, in the original ASM routine R_DrawSLSColumn it seems that the register bh can get to be larger than 10 because bx is assigned the shadingtable... Which SHOULD be 0, but when you move further away from the cloaked shape it gets darker, because bh is larger than 10.

I've poked around with the assembly for a while, and easily prevented the increment from happening, which makes it look the same as in this port. But I've so far been unable to figure out why the high byte of shading table has something in it when moving away from the cloaked shape.