Fix Bio-Techs being pushed into walls
PlayerIsBlocking doesn't test if the move is tilemap valid, so instead of just moving him to the opposite tile, first test if he can go there with TryWalk.

TryWalk will update obj->distance if it returns true, so we have to save the original distance in order to use it for switching obj->distance around.

After testing I was unable to get a Bio-Tech to walk into a wall, but there might still be a rare chance of it happening.