Fix some demo issues
The original US_RndT assembly code increments rndindex first and then indexes rndtable with it; a trivial difference except in demos where it's enough to throw off the timing and desync it.

Also there's a much longer blackscreen than there should be when the demo starts, probably caused by trying to fizzlein a faded screen.