Fix alignment issues when loading in VGAHEAD
The 3 byte starts in VGAHEAD are an annoyance that are much simpler to deal with by just reading in the entire thing 3 bytes at a time.

Since the whole VGAGRAPH is cached in on startup now, we don't need grstarts or grhuffman hanging about in memory, so they can both be freed when we're done.