Implement resolutions higher than 320x200
The 3D renderer directly calculates a higher res viewwidth/viewheight for drawing to the viewport, whereas the 2D renderer still works with the base resolution, but everything drawn gets scaled up via the integral scaling factor screen.scale. As a general rule of thumb, use the normal view variables when dealing with the 3D renderer, and the baseview variables when dealing with the 2D renderer.

Resolutions must be multiples of 320x200 or 320x240! (Multiples of 320x240 work, but are not fully implemented yet.)