Fix resolutions with a height multiple of 240
This method treats the 2D renderer as a centered 320x200 canvas, while the 3D renderer uses the entire screen to draw the 2D graphics (and the viewport for the 3D area as usual). The new screen buffer offset MUST be 0 when entering the 3D renderer! Failure to call VW_SetBufferOffset(0); before entering it would write pixels outside the screen buffer. To prevent this, the game will bomb out letting the player know the offset is non-zero. Nevertheless, there are times when 3D rendering functions like DrawPlayScreen are called from within a 2D rendering function (such as InputFloor). Be sure to reset the buffer offset before calling any functions like this!