Automatically setup display via defaults or saved variables
This approach necessitates calling ReadConfig before calling the id manager startup routines, so a couple things for the sound manager had to be moved around (though they really belonged where I've put them now).

Resolutions can also be manually set via command line parameters. This setting will override the automatic defaults as well as the saved values in the config file. Some kind of menu screen for the display would be great to have too.

Also added the ability to use a software fallback for the renderer if a hardware accelerated one is unavailable.