Fix viewheight calculation
The way this was being calculated scaled up the odd bits in the result of (size << 4) * HEIGHTRATIO, which caused viewheight to be slightly larger than it should have been. So now we mask off the 1st bit to force the base height to be even before the scaling happens.

This will result in an odd baseviewheight, which is fine because it'll get scaled up again when passed to any drawing routines (e.g. VW_Bar).