Don't show pushwall tiles with no walls on radar
On maps 3 & 4 there are pushwall tiles without wall tiles under them (so you can push a wall into that spot and push it again), but because of the way DrawRadar checks TT_TRAVELED and tilemap, it doesn't account for this, so it draws a "seen" spot on the overhead even though the player hasn't seen it yet.

As far as I can tell however, there's no need for these two lines to exist, because TT_TRAVELLED will be set on any tiles the player sees anyway.