Fix mouse/joystick detection
Since ReadConfig gets called before IN_Startup now, the latter overrides what the former sets, so now we set the mouseenabled/joystickenabled variables to true by default or read in the saved values, disabling them only if the hardware isn't found.

I don't think the SDL library even has a way of checking if there's a mouse present, but it's probably a good idea to keep the MousePresent check for systems that do.