Add menu for defining special keys
This is some terrible code that was hacked up to work with the rigid menu routines which made a better design too difficult to implement without rewriting the whole thing. I can't stress it enough, this code is DISGUSTING; it'll definitely be revisited later for some much needed refactoring, but in the meantime, at least you can redefine these keys.

Note that a WASD scheme is assumed to be in use. The arrow keys still work the same way as before, so don't try to redefine them.

As usual, these changes can be disabled by compiling with the CLASSIC_MENU define enabled. Note that this won't remove all the code from the EXE, it just prevents certain stuff from being drawn/selected. This code is already a terrible mess without throwing excessive #ifdef littering into the spaghetti!