Fix music playing too fast
I'm not entirely sure of the cause, but it's possible for the frequency to be incorrect for some, making the music play way too fast. So we allow it to change if required.

The custom samplerate is also capped to prevent the user from passing 0 Hz (if they can, it's just a question of when).