Fix overwriting reserved tempobjs as the player
This happens because obj->tempobj is set as the player when obj dies, which caused large canister aliens to call RemoveObj (player). So now we dedicate another objstruct pointer to hold the attacker's address.

It definitely seems like a waste to store something in objstruct that is only used in one place, but I can't think of a good way around it right now. At least the pointer could come in useful for something else in the future.