Fix pushwalls on overhead
My initial attempt to fix pushwall tiles with no wall tile under them showing up on the overhead (without zooming in) actually went too far; now pushwalls don't show up at all when zooming in the radar. So we do need to mark wall tiles with TT_TRAVELED (obviously...).

However, there's still the problem of pushwall tiles with no wall tile under them not being shown when zooming in (because we removed the TT_TRAVELED flag for them), so instead we update the travel table if a pushwall crossed into an empty pushwall tile.

I also decided on a straightforward way of removing the moved pushwall tiles from the overhead map when loading a saved game/level: since we can't delete them from mapsegs directly, we OR a new flag (TT_PWALLMOVED) onto the spot in TravelTable, and remove it if the block crossed into an empty pushwall tile (so that it can be drawn again). The new flag was a necessity; we can't just remove the TT_TRAVELED flag because spotvis will immediately add it back when you look at the wall.

Hopefully this will finally fix the issues with pushwalls on the overhead map.