Fix wall/sprite scaling in Text Presenter
Sprites were easy to fix, we just need to multiply the parameters by screen.scale. Walls are more tricky however because ScalePost assumes the wall will be centered vertically on the screen. If the scaled height is larger than the centered position it will simply clip the former, so we can't use that. In the end I decided the simplest approach was to convert the wall image format (wall columns are stored row major) and draw it with VW_MemToScreen.