Fix Text Presenter cursor position
I missed that TP_WrapText assigns scan_x as cur_x and then checks if it goes past the drawing bounds. This is why it didn't matter (this time) if cur_x wraps around as an unsigned int. However we can just manually set it to be 1 pixel greater than the right bound, which effectively does the same thing.

Hopefully this will fix all issues with the text presenter not wrapping text properly. LoseScreen has also been tested and it still works properly.