Fix barrier anims playing one frame after turning off
The original code used temp3 as a fake pointer to the anim struct, which is why it gets set to 0 in TurnPostOff. But now we need to explicitly set the animtype to at_NONE in order to prevent AnimateOfsObj from being executed after turning off the barrier.

There may be some redundant temp3 usage since changing how the animstruct is accessed, and it's possible that animtype needs setting elsewhere for other objects.