Fix linked patrolling Bio-techs being given informant values
This happens because they called CheckForSpecialTile after giving the special informant values. But if the Bio-tech happens to have a link, it'll remove the FL_INFORMANT flag and make him behave like an enemy with 1 hp and a seek tile. The really strange thing is that in the DOS version, the actor is still able to move with T_Chase, despite having the FL_HAS_TOKENS flag (which is the same as FL_BARRIER_DAMAGE) which should cause an early return in T_Chase. When this bug happens in this code he definitely can't chase, so I've no idea why this happens. Maybe it's some Borland weirdness when defining FL_BARRIER_DAMAGE as FL_HAS_TOKENS?

In any case, it was definitely a pretty big oversight, and this rearrangement should fix it.