Don't modify the original grsegs message chunk
This behavior wasn't necessarily dangerous, since subsequent calls to TerminateString would locate the null terminator that replaced the end marker, but it's pretty sus, and there's still the main potential issue of someone forgetting to include the end marker, since strstr() will simply keep on scanning until it finds one or happens to bump into a random 0.

In any case, it's probably wise to leave the original chunk intact and print a null terminated copy instead.