Cleanup filename handling
This was quite a mess, and was causing format-truncate warnings, so it's better to just copy the entire filename once and leave it at that. The missing file checking done in CheckForEpisodes is much more robust now too.

Also changed all the file names to lowercase, which should be easier to deal with on case sensitive systems.