A bit of cleanup
It seems like the plasma detonator was once a usable weapon, but I don't know what wp_SPACER was for. It can just be deleted and weapon loops can use < NUMWEAPONS for their range, which is much better.

Also deleted all the old dirscan table since it's no longer used.