Fix crashes when using walls on the left/top edges of the map
My fist attempt to fix this only worked for the rightmost/bottom edges.

Not sure what I thought I was doing, but player->tilex + x being 0 is perfectly valid anyway. I guess I thought that the tile members being unsigned would make them overflow to the max and the other check would then catch them, but that's not how it works. So now we make sure that the player's tile position is within the range of 2-61.